ca marche

+++ srcs/vector/fn_resize.hpp
@@ -2,14 +2,14 @@ void	fn_resize() {
 	ft::vector<int> myvector;
 
 	// set some initial content:
-	for (int i=1;i<10;i++) myvector.push_back(i);
+	for (size_t i=1;i<10;i++) myvector.push_back(i);
 
 	myvector.resize(5);
 	myvector.resize(8,100);
 	myvector.resize(12);
 
 	std::cout << "myvector contains:";
-	for (int i=0;i<myvector.size();i++)
+	for (size_t i=0;i<myvector.size();i++)
 		std::cout << ' ' << myvector[i];
 	std::cout << '\n';
 }
